Added In-App Browser.
Big update: now when a cell is tapped, the shot is displayed in a
UIWebView so there's no more that jumping to Safari and then back to
the app.

I made it pretty dirty and haven't used any of those modal browsers you
find on GitHub: just a WebView. You can find all the browser stuff in
the ViewController.m.
Now 'didSelectedRowAtindexPath' is a bit messy because I'm not very
familiar with this projects so I have created an NSURL from aShot.url
and then converted it into a NSString to load load it… maybe there's a
smarter way.